mutations

+++ src/main/java/com/graphql/spring/jpa/repositories/StoreRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
     List<Store> findAll();
 }
